# Support sharing ModuleInfoCache between pipeline runs