--- a/src/Juvix/Compiler/Pipeline/ImportParents.hs
+++ b/src/Juvix/Compiler/Pipeline/ImportParents.hs
@@ -3,23 +3,9 @@ module Juvix.Compiler.Pipeline.ImportParents where
 import Juvix.Compiler.Concrete.Data
 import Juvix.Prelude.Base
 
-data ImportParents = ImportParents
-  { _importParents :: [TopModulePath],
-    _importParentsSeenModules :: HashSet TopModulePath
+newtype ImportParents = ImportParents
+  { _importParents :: [TopModulePath]
   }
+  deriving newtype (Semigroup, Monoid)
 
 makeLenses ''ImportParents
-
-instance Semigroup ImportParents where
-  s1 <> s2 =
-    ImportParents
-      { _importParents = s1 ^. importParents <> s2 ^. importParents,
-        _importParentsSeenModules = s1 ^. importParentsSeenModules <> s2 ^. importParentsSeenModules
-      }
-
-instance Monoid ImportParents where
-  mempty =
-    ImportParents
-      { _importParents = mempty,
-        _importParentsSeenModules = mempty
-      }